Use 32-bit CharacterId in guild code
Change CharacterId from Integer to Long across guild-related modules and update protocol parsing to use 32-bit IDs. Protocol.bas now reads CharacterId with ReadInt32 (was ReadInt16). Function signatures updated in clsClan, modDatabase and modGuilds (AceptarNuevoMiembro, DetallesSolicitudAspirante, NuevoAspirante, RetirarAspirante, GetUserGuildMemberDatabase, GetUserGuildPedidosDatabase, SaveUserGuildPedidos, a_AceptarAspirante, SaveUserGuildPedidos) to accept Long. These changes fix mismatched id sizes and allow larger character IDs.

+++ Codigo/Protocol.bas
@@ -3627,7 +3627,7 @@ Private Sub HandleGuildAcceptNewMember(ByVal UserIndex As Integer)
         Dim tUser    As t_UserReference
         Dim CharacterId As Integer
         username = reader.ReadString8()
-        CharacterId = reader.ReadInt16()
+        CharacterId = reader.ReadInt32()
         tUser = NameIndex(username)
         If IsValidUserRef(tUser) Then
             If Not modGuilds.a_AceptarAspirante(UserIndex, CharacterId, username, errorStr) Then
@@ -3692,7 +3692,7 @@ Private Sub HandleGuildKickMember(ByVal UserIndex As Integer)
         Dim GuildIndex As Integer
         Dim CharacterId As Integer
         username = reader.ReadString8()
-        CharacterId = reader.ReadInt16()
+        CharacterId = reader.ReadInt32()
         GuildIndex = modGuilds.m_EcharMiembroDeClan(UserIndex, CharacterId)
         If GuildIndex > 0 Then
             Dim expulsado As t_UserReference
